+++ CHANGELOG.markdown
@@ -5,6 +5,7 @@
 * New `RULES` pragmas now avoid performing work for features you aren't using in a given `Traversal`, such as the index, etc.
 * Factored out a common `reversed` definition from all the various forms of it around the library and placed it in `Control.Lens.Iso`.
 * Added `binary, `octal`, `decimal` and `hex` to `Numeric.Lens`.
+* Added `\_Void` to Control.Lens.Void
 
 3.8.5
 -----